feat: merge prototype-app into sdk-app as dual-mode application
Add design mode to sdk-app with URL-based switching (/design/* routes).
Prototype code moved to sdk-app/src/design/, prototype-app deleted.
TopBar slimmed to title + mode switcher, env config moved to new Footer.
