+++ package.json
@@ -77,11 +77,7 @@
     "sdk-app:local": "npx tsx sdk-app/scripts/start.ts dev --env=local",
     "sdk-app-prod": "npx tsx sdk-app/scripts/start.ts prod --env=demo",
     "sdk-app:setup": "npx tsx sdk-app/scripts/setup.ts",
-    "sdk-app:analyze-props": "npx tsx sdk-app/scripts/analyze-component-props.ts",
-    "prototype-app": "npx tsx prototype-app/scripts/start.ts --env=demo",
-    "prototype-app:demo": "npx tsx prototype-app/scripts/start.ts --env=demo",
-    "prototype-app:staging": "npx tsx prototype-app/scripts/start.ts --env=staging",
-    "prototype-app:local": "npx tsx prototype-app/scripts/start.ts --env=local"
+    "sdk-app:analyze-props": "npx tsx sdk-app/scripts/analyze-component-props.ts"
   },
   "devDependencies": {
     "@commitlint/cli": "^20.5.0",
