--- prototype-app/src/App.module.scss
@@ -1,151 +1,0 @@
-:root {
-  --color-bg: #ffffff;
-  --color-bg-sidebar: #f8f9fa;
-  --color-bg-topbar: #ffffff;
-  --color-text: #1a1a2e;
-  --color-text-topbar: #131313;
-  --color-text-muted: #6c757d;
-  --color-border: #dee2e6;
-  --color-active: #0d6efd;
-  --color-active-bg: #e7f1ff;
-  --color-hover-bg: #f0f0f0;
-  --color-success: #198754;
-  --color-warning: #ffc107;
-  --color-danger: #dc3545;
-  --color-badge-bg: #e9ecef;
-}
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  font-family:
-    -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
-  font-size: 0.875rem;
-  color: var(--color-text);
-  background: var(--color-bg);
-}
-
-.layout {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-}
-
-.nav {
-  display: flex;
-  align-items: center;
-  padding: 0.75rem 1.5rem;
-  background: var(--color-bg-topbar);
-  border-bottom: 1px solid var(--color-border);
-  box-shadow: 0px 2px 5px 0px rgba(0, 0, 0, 0.05);
-  position: relative;
-  z-index: 10;
-}
-
-.title {
-  font-weight: 600;
-  font-size: 0.875rem;
-  color: var(--color-text-topbar);
-  text-decoration: none;
-
-  span {
-    font-weight: 400;
-    color: var(--color-text-muted);
-    margin-left: 0.25rem;
-    font-size: 1rem;
-  }
-}
-
-.navSpacer {
-  flex: 1;
-}
-
-.navInfo {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.navDetail {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-size: 0.75rem;
-  color: var(--color-text-muted);
-  font-family: monospace;
-  max-width: 16rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.navLabel {
-  font-family:
-    -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
-  font-size: 0.625rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.03rem;
-  opacity: 0.6;
-}
-
-.tokenStatus {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-size: 0.75rem;
-  color: var(--color-text-muted);
-}
-
-.tokenDot {
-  width: 0.5rem;
-  height: 0.5rem;
-  border-radius: 50%;
-}
-
-.tokenDot--valid {
-  background: var(--color-success);
-}
-
-.tokenDot--expired {
-  background: var(--color-danger);
-}
-
-.tokenDot--checking {
-  background: var(--color-warning);
-  animation: pulse 1s infinite;
-}
-
-.tokenDot--unknown {
-  background: var(--color-text-muted);
-}
-
-.tokenDot--none {
-  background: var(--color-text-muted);
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.4;
-  }
-}
-
-.body {
-  display: flex;
-  flex: 1;
-  overflow: hidden;
-}
-
-.main {
-  flex: 1;
-  overflow-y: auto;
-  padding: 2rem;
-}
