--- prototype-app/src/Sidebar.module.scss
@@ -1,91 +1,0 @@
-.root {
-  width: 16.25rem;
-  background: var(--color-bg-sidebar);
-  border-right: 0.0625rem solid var(--color-border);
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  flex-shrink: 0;
-}
-
-.search {
-  padding: 0.75rem;
-  border-bottom: 0.0625rem solid var(--color-border);
-}
-
-.list {
-  flex: 1;
-  overflow-y: auto;
-  padding: 0.5rem 0;
-}
-
-.category {
-  margin-bottom: 0.25rem;
-}
-
-.categoryHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.5rem 1rem;
-  cursor: pointer;
-  font-weight: 600;
-  font-size: 0.75rem;
-  text-transform: uppercase;
-  letter-spacing: 0.03125rem;
-  color: var(--color-text-muted);
-  user-select: none;
-
-  &:hover {
-    color: var(--color-text);
-  }
-}
-
-.categoryCount {
-  background: var(--color-badge-bg);
-  padding: 0.0625rem 0.375rem;
-  border-radius: 0.625rem;
-  font-size: 0.6875rem;
-  font-weight: 500;
-}
-
-.categoryArrow {
-  display: inline-flex;
-  font-size: 0.625rem;
-  transition: transform 0.15s;
-  margin-right: 0.375rem;
-  transform: rotate(90deg);
-  position: relative;
-  top: toRem(5);
-}
-
-.categoryArrowCollapsed {
-  transform: rotate(0deg);
-}
-
-.items {
-  list-style: none;
-}
-
-.item a {
-  display: block;
-  padding: 0.375rem 1rem 0.375rem 1.75rem;
-  color: var(--color-text);
-  text-decoration: none;
-  font-size: 0.8125rem;
-  border-left: 0.1875rem solid transparent;
-  transition:
-    background 0.1s,
-    border-color 0.1s;
-
-  &:hover {
-    background: var(--color-hover-bg);
-  }
-
-  &:global(.active) {
-    background: var(--color-active-bg);
-    border-left-color: var(--color-active);
-    color: var(--color-active);
-    font-weight: 500;
-  }
-}
