--- prototype-app/src/TokenBanner.module.scss
@@ -1,17 +1,0 @@
-.banner {
-  background: #fff3cd;
-  border-bottom: 1px solid #ffecb5;
-  padding: 0.625rem 1.5rem;
-}
-
-.content {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  font-size: 0.8125rem;
-}
-
-.message {
-  color: #664d03;
-  font-weight: 500;
-}
