--- prototype-app/src/prototypes/sample-flow/SampleFlow.module.scss
@@ -1,39 +1,0 @@
-.layout {
-  max-width: toRem(640);
-}
-
-.nav {
-  display: flex;
-  gap: toRem(4);
-  margin-bottom: toRem(24);
-  border-bottom: 1px solid #dee2e6;
-}
-
-.step,
-.stepActive {
-  padding: toRem(8) toRem(16);
-  font-size: toRem(14);
-  text-decoration: none;
-  color: #6c757d;
-  border-bottom: 2px solid transparent;
-  margin-bottom: -1px;
-  transition:
-    color 0.15s,
-    border-color 0.15s;
-
-  &:hover {
-    color: #212529;
-  }
-}
-
-.stepActive {
-  color: #0d6efd;
-  border-bottom-color: #0d6efd;
-  font-weight: 500;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: toRem(16);
-}
