--- prototype-app/src/scss.d.ts
@@ -1,1 +1,0 @@
-declare module '*.scss'
