--- prototype-app/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": ["./src", "./scripts", "../src"],
-  "compilerOptions": {
-    "ignoreDeprecations": "6.0",
-    "rootDir": "..",
-    "baseUrl": "..",
-    "paths": {
-      "@/*": ["src/*"]
-    },
-    "noEmit": true
-  }
-}
