+++ sdk-app/src/App.tsx
@@ -1,17 +1,20 @@
 import { useState } from 'react'
 import { Outlet } from 'react-router-dom'
 import { TopBar } from './TopBar'
+import { Footer } from './Footer'
 import { Sidebar } from './Sidebar'
 import { DemoSettingsPanel } from './DemoSettingsPanel'
 import { TokenExpiredOverlay } from './TokenExpiredOverlay'
 import { useEntities } from './useEntities'
 import { useDemoManager } from './useDemoManager'
+import { useAppMode } from './useAppMode'
 
 export function App() {
   const [settingsOpen, setSettingsOpen] = useState(false)
   const [searchQuery, setSearchQuery] = useState('')
   const { entities, updateEntity, replaceEntities, resetToDefaults } = useEntities()
   const demoManager = useDemoManager()
+  const mode = useAppMode()
 
   const handleCreateNewDemo = async (demoType: string) => {
     const result = await demoManager.createNewDemo(demoType)
@@ -29,19 +32,20 @@ export function App() {
 
   return (
     <div className="app-layout">
-      <TopBar
+      <TopBar />
+      <div className="app-body">
+        <Sidebar mode={mode} searchQuery={searchQuery} onSearchChange={setSearchQuery} />
+        <main className="main-content">
+          <Outlet context={{ entities }} />
+        </main>
+      </div>
+      <Footer
         companyId={entities.companyId}
         tokenStatus={demoManager.tokenStatus}
         onOpenSettings={() => {
           setSettingsOpen(true)
         }}
       />
-      <div className="app-body">
-        <Sidebar searchQuery={searchQuery} onSearchChange={setSearchQuery} />
-        <main className="main-content">
-          <Outlet context={{ entities }} />
-        </main>
-      </div>
       <DemoSettingsPanel
         isOpen={settingsOpen}
         onClose={() => {
