+++ sdk-app/src/ModeSwitcher.module.scss
@@ -1,0 +1,33 @@
+.root {
+  display: inline-flex;
+  border: 1px solid var(--color-border);
+  border-radius: 0.375rem;
+  overflow: hidden;
+}
+
+.option {
+  padding: 0.25rem 0.75rem;
+  font-size: 0.75rem;
+  font-weight: 500;
+  text-decoration: none;
+  color: var(--color-text-muted);
+  background: transparent;
+  transition:
+    background 0.15s,
+    color 0.15s;
+
+  &:hover {
+    background: var(--color-hover-bg);
+    color: var(--color-text);
+  }
+}
+
+.active {
+  background: var(--color-active);
+  color: #fff;
+
+  &:hover {
+    background: var(--color-active);
+    color: #fff;
+  }
+}
