+++ sdk-app/src/Sidebar.module.scss
@@ -65,13 +65,17 @@
 }
 
 .categoryArrow {
+  display: inline-flex;
   font-size: 0.625rem;
   transition: transform 0.15s;
   margin-right: 0.375rem;
+  transform: rotate(90deg);
+  position: relative;
+  top: toRem(5);
 }
 
 .categoryArrowCollapsed {
-  transform: rotate(-90deg);
+  transform: rotate(0deg);
 }
 
 .items {
