+++ sdk-app/src/TopBar.module.scss
@@ -2,141 +2,23 @@
   display: flex;
   align-items: center;
   height: 3rem;
-  background: var(--color-bg-topbar);
-  color: var(--color-text-topbar);
+  background: #ffffff;
+  color: var(--color-text);
   padding: 0 1rem;
   gap: 1rem;
   flex-shrink: 0;
   font-size: 0.8125rem;
   z-index: 10;
+  border-bottom: 1px solid var(--color-border);
+  box-shadow: 0px 2px 5px 0px rgba(0, 0, 0, 0.05);
 }
 
 .title {
   font-weight: 600;
   font-size: 0.9375rem;
-  color: #fff;
-  margin-right: 0.5rem;
-}
-
-.divider {
-  width: 0.0625rem;
-  height: 1.5rem;
-  background: rgba(255, 255, 255, 0.2);
-}
-
-.badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.375rem;
-  padding: 0.1875rem 0.625rem;
-  border-radius: 0.25rem;
-  font-size: 0.6875rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.03125rem;
-}
-
-.badgeLabel {
-  font-weight: 400;
-  opacity: 0.7;
-  font-size: 0.625rem;
-}
-
-.badgeEnv {
-  background: rgba(13, 110, 253, 0.3);
-  color: #8bb9fe;
-}
-
-.badgeBuild {
-  background: rgba(25, 135, 84, 0.3);
-  color: #7dd3a8;
-}
-
-.badgeDemoType {
-  background: rgba(255, 193, 7, 0.25);
-  color: #f0d060;
-}
-
-.company {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-family: monospace;
-  font-size: 0.75rem;
-  opacity: 0.8;
-  max-width: 16.25rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-
-  .badgeLabel {
-    font-family:
-      -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
-    font-size: 0.625rem;
-    font-weight: 400;
-    text-transform: uppercase;
-    letter-spacing: 0.03125rem;
-  }
-}
-
-.tokenStatus {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-size: 0.75rem;
-}
-
-.tokenDot {
-  width: 0.5rem;
-  height: 0.5rem;
-  border-radius: 50%;
-}
-
-.tokenDotValid {
-  background: var(--color-success);
-}
-
-.tokenDotExpired {
-  background: var(--color-danger);
-}
-
-.tokenDotChecking {
-  background: var(--color-warning);
-  animation: pulse 1s infinite;
-}
-
-.tokenDotUnknown {
-  background: var(--color-text-muted);
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.4;
-  }
+  color: var(--color-text);
 }
 
 .spacer {
   flex: 1;
-}
-
-.gear {
-  background: none;
-  border: 0.0625rem solid rgba(255, 255, 255, 0.2);
-  color: var(--color-text-topbar);
-  cursor: pointer;
-  padding: 0.375rem 0.625rem;
-  border-radius: 0.375rem;
-  font-size: 0.8125rem;
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  transition: background 0.15s;
-
-  &:hover {
-    background: rgba(255, 255, 255, 0.1);
-  }
 }
