+++ sdk-app/src/app.scss
@@ -1,9 +1,9 @@
 :root {
   --color-bg: #ffffff;
   --color-bg-sidebar: #f8f9fa;
-  --color-bg-topbar: #1a1a2e;
+  --color-bg-topbar: #ffffff;
   --color-text: #1a1a2e;
-  --color-text-topbar: #e0e0e0;
+  --color-text-topbar: #1a1a2e;
   --color-text-muted: #6c757d;
   --color-border: #dee2e6;
   --color-active: #0d6efd;
