+++ sdk-app/src/design/DesignLayout.module.scss
@@ -1,0 +1,3 @@
+.root {
+  padding: 2rem;
+}
